AMBARI-1937. Ambari-web installer wizard doesn't work in test mode. (jaimin)
git-svn-id: https://svn.apache.org/repos/asf/incubator/ambari/trunk@1468272 13f79535-47bb-0310-9956-ffa450edef68

+++ ambari-web/app/controllers/wizard/step8_controller.js
@@ -867,15 +867,13 @@ App.WizardStep8Controller = Em.Controller.extend({
       default:
         break;
     }
-
     // delete any existing clusters to start from a clean slate
     // before creating a new cluster in install wizard
     // TODO: modify for multi-cluster support
-    if (this.get('content.controllerName') == 'installerController') {
+    if (this.get('content.controllerName') == 'installerController' && (!App.testMode)) {
       var clusterNames = this.getExistingClusterNames();
       this.deleteClusters(clusterNames);
     }
-
     this.createCluster();
     this.createSelectedServices();
     this.createConfigurations();
